Controllers(Book, Member, Borrow) added + Id fix

+++ Library/Library_Common/Models/Borrow.cs
@@ -8,9 +8,9 @@ namespace Library_Common.Models
 {
     public class Borrow
     {
-        public int Id { get; set; }
-        public int BookId { get; set; }
-        public int MemberId { get; set; }
+        public long Id { get; set; }
+        public long BookId { get; set; }
+        public long MemberId { get; set; }
         public DateTime BorrowDate { get; set; }
         public DateTime ReturnDate { get; set; }
